# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    All endpoints (except `/health` and `/auth-init`) require a bearer token
    via the `Authorization: Bearer <token>` header or `?token=` query param.

    **Two-tier auth (optional):** When `WRITE_TOKEN` is configured, mutating
    operations (POST, PATCH, PUT, DELETE) require the write token. Read
    operations (GET) accept either the read token (`AUTH_TOKEN`) or write
    token. If `WRITE_TOKEN` is not set, all operations accept `AUTH_TOKEN`
    (backwards compatible).

    The `/auth-init` endpoint returns only the read token — the write token
    is never exposed by any endpoint.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
      Talkgroups and units belong to the system level because they are
      shared across all sites on the same P25 network.

    - **Site** (`site_id`): a recording point within a system. Each
...
      specific site.

...
    ```
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /units:
    get:
      operationId: listUnits
      summary: List radio units
      description: |
        Returns radio units with optional filters. When searching, results
        include relevance_score (100=exact, 50=prefix, 10=contains).
      tags: [units]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID
          schema:
            type: string
        - name: search
          in: query
          description: Search by alpha_tag or unit_id
          schema:
            type: string
        - name: active_within
          in: query
          description: |
            Only return units active within this many minutes. Replaces the
            need for a separate "active units" endpoint when combined with
            other filters.
          schema:
            type: integer
        - name: talkgroup
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple)
          schema:
            type: string
            example: "9178,5344"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/sortDir"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /units/{id}:
    get:
      operationId: getUnit
      summary: Get a unit
      description: |
        Returns a single unit. Accepts composite `system_id:unit_id` (e.g.,
        "1:1234567") or plain `unit_id`. Returns 409 if ambiguous.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateUnit
      summary: Update unit metadata
      description: >-
        Updates mutable unit fields (alpha_tag, alpha_tag_source).
        When CSV_WRITEBACK is enabled and TR_DIR is configured with a unitTagsFile,
        alpha_tag changes are also written back to trunk-recorder's unit CSV file on disk.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"
  /units/{id}/calls:
    get:
      operationId: listUnitCalls
      summary: List calls for a unit
      description: |
        Returns calls that include transmissions from a specific unit.
        Accepts composite `system_id:unit_id` or plain `unit_id`.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"
  /units/{id}/events:
    get:
      operationId: listUnitEvents
      summary: List unit events
      description: |
        Returns events (affiliations, registrations, calls, etc.) for a
        specific unit. Includes embedded talkgroup and unit display names.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
        - name: type
          in: query
          description: Filter by event type
          schema:
            $ref: "#/components/schemas/EventType"
        - name: talkgroup
          in: query
          description: Filter by talkgroup ID
          schema:
            type: integer
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitEventListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Unit Events (system-wide)
  # ----------------------------------------------------------
...
  /calls:
    get:
      operationId: listCalls
      summary: List calls
      description: |
        Returns call recordings with comprehensive filters. This is the
        primary endpoint for querying call history.

        **Common patterns:**
        - Live feed: `?sort=-stop_time&deduplicate=true&limit=50`
        - Talkgroup history: `?tgid=9178&sysid=348&start_time=...&end_time=...`
        - Unit activity: `?unit_id=924003`
        - Emergency calls: `?emergency=true&sort=-start_time`
      tags: [calls]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID (comma-separated for multiple, e.g., "348,34D"). Alias "sysids" also accepted.
          schema:
            type: string
        - name: system_id
          in: query
          description: Filter by system database ID (comma-separated for multiple). Alias "systems" also accepted.
          schema:
            type: string
            example: "1,2"
        - name: site_id
          in: query
          description: Filter by recording site ID (comma-separated for multiple). Alias "sites" also accepted.
          schema:
            type: string
            example: "1,3"
        - name: tgid
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple). Alias "tgids" also accepted.
          schema:
            type: string
            example: "9178,5344"
        - name: unit_id
          in: query
          description: Filter by unit radio ID (comma-separated for multiple). Aliases "units", "unit_ids" also accepted.
          schema:
            type: string
            example: "924003,104"
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - name: encrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
        - name: deduplicate
          in: query
          description: |
            When true, returns one call per call_group (best recording).
            Useful for feed views where simulcast duplicates are unwanted.
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - name: sort
          in: query
          description: |
            Sort field. Prefix with `-` for descending.
            Allowed: start_time, stop_time, duration, tgid, freq
          schema:
            type: string
            default: "-start_time"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
...
    unitId:
      name: id
      in: path
      required: true
      description: "Unit ID: `system_id:unit_id` (e.g., `1:924003`) or plain `unit_id`"
      schema:
        type: string
        example: "1:924003"
    startTime:
      name: start_time
      in: query
      description: Start of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    endTime:
      name: end_time
      in: query
      description: End of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
    sort:
      name: sort
      in: query
      description: "Sort field. Prefix with `-` for descending (e.g., `-last_seen`)"
      schema:
        type: string
        default: "alpha_tag"
    sortDir:
      name: sort_dir
      in: query
      description: Sort direction (ignored if sort field uses `-` prefix)
      schema:
        type: string
        enum: [asc, desc]
        default: asc

  # ----------------------------------------------------------
  # Reusable Responses
  # ----------------------------------------------------------
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Ambiguous:
      description: |
        Ambiguous ID — the plain ID exists in multiple systems.
        Response includes the list of matching systems to help the
        client disambiguate.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AmbiguousError"
...
    Forbidden:
      description: Forbidden — valid read token but write token required for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    EventType:
      type: string
      enum: [on, off, join, call, end, data, ans_req, location, ackresp]
      description: |
        Unit event types from trunk-recorder MQTT:
        - **on**: unit registration (radio powered on / entering system)
        - **off**: unit deregistration (radio powered off / leaving system)
        - **join**: talkgroup affiliation (unit joins a talkgroup)
        - **call**: channel grant (voice call participation)
        - **end**: transmission end (per-unit end-of-transmission report)
        - **data**: data grant
        - **ans_req**: answer request
        - **location**: unit location update
        - **ackresp**: acknowledge response
    CallState:
      type: string
      enum: [monitoring, recording, stopped, completed]
      description: |
        Call recording state (mapped from trunk-recorder integer values):
        - **monitoring**: control channel activity seen, no audio capture yet
        - **recording**: actively capturing audio
        - **stopped**: recording ended, finalizing
        - **completed**: audio file written, ready for access
    MonState:
      type: string
      enum: [unspecified, active, duplicate]
      description: |
        Call monitor state (from trunk-recorder):
        - **unspecified**: default / not specifically monitored
        - **active**: actively monitored
        - **duplicate**: duplicate call on another recorder
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
    AmbiguousError:
      type: object
      required: [code, error, matches]
      properties:
        code:
          type: string
          example: ambiguous_id
        error:
          type: string
          example: "Ambiguous ID 9178: exists in multiple systems"
        matches:
          type: array
          description: |
            Systems where this ID was found. Includes system_id so the
            client can construct the composite ID to disambiguate.
          items:
            type: object
            properties:
              system_id:
                type: integer
                description: Database system ID — use this to form the composite ID
                example: 1
              system_name:
                type: string
                description: System display name
                example: "Butler/Warren P25"
              sysid:
                type: string
                description: P25 SYSID (for display context)
                example: "348"

    # ========== Core Domain Models ==========
...
    Unit:
      type: object
      description: A radio unit (mobile or portable radio)
      required: [system_id, unit_id]
      properties:
        system_id:
          type: integer
          description: Database system ID this unit belongs to
          example: 1
        system_name:
          type: string
          description: System display name (embedded to avoid lookup)
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"
        unit_id:
          type: integer
          example: 1234567
        alpha_tag:
          type: string
          example: Engine 1
        alpha_tag_source:
          type: string
          description: Source of the alpha tag (e.g., radioreference, manual)
          example: radioreference
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        # Last event context (embedded)
        last_event_type:
          $ref: "#/components/schemas/EventType"
        last_event_time:
          type: string
          format: date-time
        last_event_tgid:
          type: integer
          example: 9178
        last_event_tg_tag:
          type: string
          description: Alpha tag of the talkgroup from last event
          example: "09-8L Main"
        call_count:
          type: integer
          description: "Number of calls within the requested time window (only present on /talkgroups/{id}/units)"
          example: 42
        # Search field
        relevance_score:
          type: integer
          description: "Search relevance: 100=exact, 50=prefix, 10=contains. Only present in search results."
          example: 100
    Call:
      type: object
      description: A recorded radio call/transmission
      required: [call_id, system_id, tgid, start_time]
      properties:
        # Identifiers
        call_id:
          type: integer
          description: Unique database ID for this recording
          example: 48531
        call_group_id:
          type: integer
          description: |
            Call group this recording belongs to. Multiple recordings of the
            same transmission (from different sites) share a call_group_id.
            Derived from (system_id, tgid, start_unix).
          example: 1

        # System context (embedded to avoid lookup)
        system_id:
          type: integer
          description: Logical radio system this call belongs to
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"

        # Site context (which recorder captured this)
        site_id:
          type: integer
          description: Recording site that captured this call
          example: 1
        site_short_name:
          type: string
          description: TR sys_name of the capturing site (e.g., "butco", "warco")
          example: butco

        # Talkgroup context (embedded to avoid lookup)
        tgid:
          type: integer
          example: 9178
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch
        tg_tag:
          type: string
          description: Functional category tag (e.g., Law Dispatch, Fire Tac)
          example: Law Dispatch
        tg_group:
          type: string
          description: Broad organizational grouping
          example: "Butler County (09) Law"

        # Timing
        start_time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        stop_time:
          type: string
          format: date-time
          nullable: true
          description: "Null for active (in-progress) calls"
          example: "2024-01-15T10:30:45Z"
        duration:
          type: number
          description: Duration in seconds (elapsed so far for active calls)
          example: 45.5

        # Audio
        audio_url:
          type: string
          nullable: true
          description: "Relative URL to audio endpoint. Null for active calls."
          example: /api/v1/calls/48531/audio
        audio_type:
          type: string
          description: "Audio encoding type from trunk-recorder (e.g., `digital`, `digital tdma`, `analog`)"
          example: digital
        audio_size:
          type: integer
          description: Audio file size in bytes
          example: 45000

        # Signal quality
        freq:
          type: integer
          description: Frequency in Hz
          example: 851012500
        freq_error:
          type: integer
          description: Frequency error in Hz
          example: 0
        signal_db:
          type: number
          nullable: true
          description: Signal strength in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -23
        noise_db:
          type: number
          nullable: true
          description: Noise floor in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -50
        error_count:
          type: integer
          description: Decode error count
          example: 2
        spike_count:
          type: integer
          description: Audio spike count
          example: 0

        # Status
        call_state:
          $ref: "#/components/schemas/CallState"
        mon_state:
          $ref: "#/components/schemas/MonState"
        emergency:
          type: boolean
          example: false
        encrypted:
          type: boolean
          example: false
        analog:
          type: boolean
          example: false
        conventional:
          type: boolean
          example: false
        phase2_tdma:
          type: boolean
          example: false
        tdma_slot:
          type: integer
          description: "TDMA slot (0=slot 1 or N/A, 1=slot 2)"
          example: 0

        # Patches
        patched_tgids:
          type: array
          description: Other talkgroup IDs patched into this call
          items:
            type: integer

        # Transmission and frequency data (JSONB columns)
        src_list:
          type: array
          description: |
            Source/transmission list from trunk-recorder's srcList.
            Each entry is a unit transmission segment within the call.
            Also available via GET /calls/{id}/transmissions.
          items:
            $ref: "#/components/schemas/CallTransmission"
        freq_list:
          type: array
          description: |
            Frequency list from trunk-recorder's freqList.
            Each entry is a frequency segment within the call.
            Also available via GET /calls/{id}/frequencies.
          items:
            $ref: "#/components/schemas/CallFrequency"
        unit_ids:
          type: array
          description: Denormalized array of distinct unit radio IDs (src values) from src_list. GIN-indexed for fast unit filtering.
          items:
            type: integer

        # Units on this call (embedded)
        units:
          type: array
          items:
            $ref: "#/components/schemas/CallUnit"

        # Transcription (denormalized from transcriptions table)
        has_transcription:
          type: boolean
          description: Whether this call has been transcribed
          example: true
        transcription_status:
          $ref: "#/components/schemas/TranscriptionStatus"
        transcription_text:
          type: string
          nullable: true
          description: Primary transcription text (null if no transcription)
          example: Engine 5 responding to the scene
        transcription_word_count:
          type: integer
          nullable: true
          example: 6

        # Extensible metadata
        metadata_json:
          type: object
          additionalProperties: true
          description: Arbitrary metadata from trunk-recorder
...
    UnitEvent:
      type: object
      description: A unit event (registration, affiliation, call activity, etc.)
      required: [id, event_type, unit_rid, unit_id, time]
      properties:
        id:
          type: integer
          example: 1
        event_type:
          $ref: "#/components/schemas/EventType"
        time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

        # System context
        system_id:
          type: integer
          description: Database system ID
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"

        # Unit context (embedded to avoid lookup)
        unit_rid:
          type: integer
          description: "P25 Radio ID (legacy name; prefer unit_id for consistency)"
          example: 1234567
        unit_id:
          type: integer
          description: "Radio unit ID (alias for unit_rid, consistent with other endpoints)"
          example: 1234567
        unit_alpha_tag:
          type: string
          description: Unit display name (embedded to avoid lookup)
          example: Engine 1

        # Talkgroup context (embedded to avoid lookup)
        tgid:
          type: integer
          example: 1001
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch

        # Internal references
        instance_id:
          type: string
          description: Trunk-recorder instance that reported this event
          example: trunk-recorder
        metadata_json:
          type: object
          additionalProperties: true
...
    UnitPatch:
      type: object
      description: Mutable unit fields. Only provided fields are updated.
      properties:
        alpha_tag:
          type: string
        alpha_tag_source:
          type: string

    # ========== Response Wrappers ==========
...
    UnitListResponse:
      type: object
      required: [units, total, limit, offset]
      properties:
        units:
          type: array
          items:
            $ref: "#/components/schemas/Unit"
        total:
          type: integer
          example: 500
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
    UnitEventListResponse:
      type: object
      required: [events, total, limit, offset]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/UnitEvent"
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
...
    CallListResponse:
      type: object
      required: [calls, total, limit, offset]
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/Call"
        total:
          type: integer
          description: Total matching results (across all pages)
          example: 2500
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
...